legacy/requirements.txt
appdirs==1.4.3
dj-database-url==0.4.1
Django==1.11.2
djangorestframework==3.7.3
django-crispy-forms==1.6.1
django-localflavor==1.5.2
django-dashboarder==0.1.1
django-cors-headers==2.1.0
celery==4.1.0
gunicorn==19.7.1
packaging==16.8
Pillow==3.3.1
psycopg2==2.7.1
pyparsing==2.2.0
pytz==2017.2
six==1.10.0
whitenoise==3.2
redis==2.10.6
